requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.1
django-adminlte-3==0.1.6
django-filter==21.1
django-pandas==0.6.6
djangorestframework==3.13.1
entrypoints==0.3
executing==0.8.2
filelock==3.4.2
frozenlist==1.2.0
geoip2==2.9.0
idna==3.3
ipykernel==6.7.0
ipython==8.0.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.1
matplotlib-inline==0.1.3
maxminddb==2.2.0
multidict==5.2.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
nest-asyncio==1.5.4
numpy==1.22.1
packaging==21.3
Paginator==0.5.1
pandas==1.4.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.4.1
plotly==5.5.0
prompt-toolkit==3.0.26
protobuf==3.19.3
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
pywin32==303
pyzmq==22.3.0
redis==4.1.3
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
tenacity==8.0.1
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.0
wcwidth==0.2.5
wrapt==1.13.3
yarl==1.7.2
